# openapi.yaml
openapi: 3.0.3
info:
  title: OOH Audio Sync API
  description: Backend APIs for the DreamBricks Audio-Video Sync Totem system.
  version: "1.0.0"
servers:
  - url: http://localhost:8001
    description: Local Node.js Server
paths:
...
  /api/stats:
    get:
      summary: Daily Statistics
      description: |
        Counters per campaign per day (time zone `STATS_TIMEZONE`), kept for 90 days, shared by every
        server on the same bucket/prefix. No personal data. `listen_seconds / listens` = average listening time.
      parameters:
        - name: days
          in: query
          schema: { type: integer, minimum: 1, maximum: 90, default: 30 }
      responses:
        '200':
          description: Oldest day first; days without activity have empty `campaigns`
          content:
            application/json:
              example:
                timezone: America/Sao_Paulo
                days:
                  - date: "2026-10-07"
                    campaigns:
                      totem1: { screens: 12, scans: 5, listeners: 4, listen_seconds: 420, listens: 4, sites: { "loja.com.br": 10, "(direto)": 2 } }
  /api/stats.csv:
    get:
      summary: Statistics Spreadsheet
      description: Same data as `/api/stats` as a `;`-separated CSV (opens in Excel), one row per day and campaign.
      parameters:
        - name: days
          in: query
          schema: { type: integer, minimum: 1, maximum: 90, default: 30 }
        - name: campaign
          in: query
          schema: { type: string }
      responses:
        '200':
          description: CSV file (download)
...
  /api/media/{filename}:
    parameters:
      - name: filename
        in: path
        required: true
        schema:
          type: string
        example: promo.mp4
    put:
      summary: Replace File Content
      description: Overwrites an existing file with the raw request body, keeping its name.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: File replaced
        '404':
          description: File not found
        '413':
          description: File exceeds the size limit
    patch:
      summary: Rename File
      description: |
        Renames a file (the name is sanitized; a video must stay a video, an audio an audio).
        Totems using the file are updated to the new name, and online totems switch video immediately.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename: { type: string, example: new-name.mp4 }
      responses:
        '200':
          description: Renamed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  filename: { type: string }
                  updated_totems: { type: array, items: { type: string } }
        '400':
          description: Invalid name or type change
        '404':
          description: File not found
        '409':
          description: Another file already has this name
    delete:
      summary: Delete File
      description: Deletes a file. Refused while any totem is configured to use it.
      responses:
        '200':
          description: Deleted
        '404':
          description: File not found
        '409':
          description: File is in use (`used_by` lists the totems)
...
  /media/{filename}:
    get:
      summary: Stream Media File
      description: Streams a video or audio file with support for HTTP Range requests.
      parameters:
        - in: path
          name: filename
          required: true
          schema:
            type: string
          description: The filename of the media to stream
      responses:
        '200':
          description: Full file content
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        '206':
          description: Partial content (Range response)
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid filename path
        '404':
          description: File not found
        '416':
          description: Range not satisfiable
...